tests: moved HDF5 tests up to Array test suite
